Update Clothing Store Database and Sats
In this project I created a table for all the items in a store, with at least 5 columns for the kind of data I'd need to store. I decided to should sell at 15 items, and use a select statement to order my items by price and show at least one statistic about the items.